runit/requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==37.0.2
dnspython==2.2.1
Flask==2.1.1
Flask-JWT-Extended==4.4.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.30
mysqlclient==2.1.1
passlib==1.7.4
protobuf==3.20.1
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
python-dotenv==0.20.0
python-runit==0.0.4
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.1
